:lipstick: :: main activity compose view 디자인
